requirements.txt
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
guess-language-spirit==0.5.3
idna==2.8
requests==2.22.0
soupsieve==1.9.3
urllib3==1.25.3
